Implement IN operator
